% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity.R
\name{entity_type}
\alias{entity_type}
\title{Entity Type, Entity}
\usage{
entity_type(entity_type_name, subconstructor = function(values) {
    values
})
}
\arguments{
\item{rule}{validation rule; <language>}
}
\value{

}
\description{
Entity Type, Entity
}
