% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniER-package.R
\docType{package}
\name{miniER-package}
\alias{miniER}
\alias{miniER-package}
\title{miniER: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Siqi Zhang \email{siqi.zhang@bayer.com}

Other contributors:
\itemize{
  \item Bayer Healthcare LLC [funder]
}

}
\keyword{internal}
