% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{register}
\alias{register}
\alias{register.miniER.entity_type}
\alias{register.miniER.relationship_type}
\alias{register.miniER.relationship}
\title{Register an Item to miniER's Workspace}
\usage{
register(item, ...)

\method{register}{miniER.entity_type}(item, ...)

\method{register}{miniER.relationship_type}(item, ...)

\method{register}{miniER.relationship}(item, ...)
}
\arguments{
\item{item}{item; <miniER.entity_type; miniER.relationship_type; miniER.relationship>}

\item{...}{dot-dot-dot}
}
\value{
item
}
\description{
Register an Item to miniER's Workspace
}
