% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relationship.R
\name{relationship}
\alias{relationship}
\alias{relationship_kind}
\alias{one_one}
\alias{one_many}
\alias{many_many}
\title{Relationship Kind, Relationship Type and Relationships}
\usage{
relationship_kind(relationship_kind_name, rule)

one_one(relationship_type_name, entity_type_1_name, entity_type_2_name)

one_many(relationship_type_name, entity_type_1_name, entity_type_2_name)

many_many(relationship_type_name, entity_type_1_name, entity_type_2_name)
}
\arguments{
\item{rule}{}
}
\value{

}
\description{
Relationship Kind, Relationship Type and Relationships
}
